+++ CHANGELOG.md
@@ -1,3 +1,12 @@
 ## 0.0.1
-Add My camera Plugin in flutter
+Add custom camera Plugin in flutter
+
+## 0.0.2
+custom camera Plugin in flutter
+
+## 0.0.3
+custom camera Plugin in flutter
+
+## 0.0.4
+custom camera Plugin in flutter